Поддержка модульности Config
