+++ lib/dapp.rb
@@ -36,8 +36,8 @@ require 'dapp/error/shellout'
 require 'dapp/exception/base'
 require 'dapp/cli'
 require 'dapp/cli/base'
-require 'dapp/config/base'
 require 'dapp/config/directive/base'
+require 'dapp/config/config'
 require 'dapp/config/error/config'
 require 'dapp/dapp/lock'
 require 'dapp/dapp/ssh_agent'
@@ -117,16 +117,14 @@ require 'dapp/dimg/cli/bp'
 require 'dapp/dimg/cli/mrproper'
 require 'dapp/dimg/cli/stage_image'
 require 'dapp/dimg/cli/cli'
-require 'dapp/dimg/config/base'
 require 'dapp/dimg/config/directive/base'
-require 'dapp/dimg/config/dimg/instance_methods'
-require 'dapp/dimg/config/dimg/validation'
-require 'dapp/dimg/config/dimg'
-require 'dapp/dimg/config/artifact_dimg'
-require 'dapp/dimg/config/dimg_group_base'
-require 'dapp/dimg/config/dimg_group_main'
-require 'dapp/dimg/config/dimg_group'
-require 'dapp/dimg/config/artifact_group'
+require 'dapp/dimg/config/directive/dimg/instance_methods'
+require 'dapp/dimg/config/directive/dimg/validation'
+require 'dapp/dimg/config/directive/dimg'
+require 'dapp/dimg/config/directive/artifact_dimg'
+require 'dapp/dimg/config/directive/dimg_group_base'
+require 'dapp/dimg/config/directive/dimg_group'
+require 'dapp/dimg/config/directive/artifact_group'
 require 'dapp/dimg/config/directive/artifact_base'
 require 'dapp/dimg/config/directive/git_artifact_local'
 require 'dapp/dimg/config/directive/git_artifact_remote'
@@ -138,6 +136,7 @@ require 'dapp/dimg/config/directive/docker/base'
 require 'dapp/dimg/config/directive/docker/dimg'
 require 'dapp/dimg/config/directive/docker/artifact'
 require 'dapp/dimg/config/directive/mount'
+require 'dapp/dimg/config/config'
 require 'dapp/dimg/dapp/command/common'
 require 'dapp/dimg/dapp/command/build'
 require 'dapp/dimg/dapp/command/bp'
@@ -160,6 +159,7 @@ require 'dapp/dimg/dapp/command/build_context/export'
 require 'dapp/dimg/dapp/command/build_context/import'
 require 'dapp/dimg/dapp/command/build_context/common'
 require 'dapp/dimg/dapp/command'
+require 'dapp/dimg/dapp/dappfile'
 require 'dapp/dimg/docker_registry'
 require 'dapp/dimg/docker_registry/base/request'
 require 'dapp/dimg/docker_registry/base/authorization'
