+++ lib/dapp/dapp.rb
@@ -92,9 +92,5 @@ module Dapp
     def stage_dapp_label
       name
     end
-
-    def dev_mode?
-      !!cli_options[:dev]
-    end
   end # Dapp
 end # Dapp
