+++ lib/dapp/dapp/dappfile.rb
@@ -9,14 +9,6 @@ module Dapp
         end
       end
 
-      def build_configs
-        @configs ||= begin
-          dimgs(dappfile_path).flatten.tap do |dimgs|
-            raise Error::Dapp, code: :no_such_dimg, data: { dimgs_patterns: dimgs_patterns.join(', ') } if dimgs.empty?
-          end
-        end
-      end
-
       def dappfile_path
         raise Error::Dapp, code: :dappfile_not_found unless (dappfile_path = search_file_upward('Dappfile'))
         dappfile_path
@@ -32,27 +24,28 @@ module Dapp
         path
       end
 
-      def dimgs(dappfile_path)
-        Dimg::Config::DimgGroupMain.new(dapp: self) do |conf|
-          begin
-            conf.instance_eval File.read(dappfile_path), dappfile_path
-          rescue SyntaxError, StandardError => e
-            backtrace = e.backtrace.find { |line| line.start_with?(dappfile_path) }
-            message = begin
-              case e
-              when NoMethodError
-                e.message =~ /`.*'/
-                "undefined method #{Regexp.last_match}"
-              when NameError then e.message[/.*(?= for)/]
-              else
-                e.message
+      def config
+        @config ||= begin
+          ::Dapp::Config::Config.new(dapp: self).tap do |config|
+            begin
+              config.instance_eval File.read(dappfile_path), dappfile_path
+              config.validate!
+            rescue SyntaxError, StandardError => e
+              backtrace = e.backtrace.find { |line| line.start_with?(dappfile_path) }
+              message = begin
+                case e
+                when NoMethodError
+                  e.message =~ /`.*'/
+                  "undefined method #{Regexp.last_match}"
+                when NameError then e.message[/.*(?= for)/]
+                else
+                  e.message
+                end
               end
+              message = "#{backtrace[/.*(?=:in)/]}: #{message}" if backtrace
+              raise Error::Dappfile, code: :incorrect, data: { error: e.class.name, message: message }
             end
-            message = "#{backtrace[/.*(?=:in)/]}: #{message}" if backtrace
-            raise Error::Dappfile, code: :incorrect, data: { error: e.class.name, message: message }
           end
-        end._dimg.select { |dimg| dimgs_patterns.any? { |pattern| dimg._name.nil? || File.fnmatch(pattern, dimg._name) } }.tap do |dimgs|
-          dimgs.each { |dimg| dimg.send(:validate!) }
         end
       end
     end # Dappfile
