+++ lib/dapp/dapp/logging/base.rb
@@ -30,6 +30,10 @@ module Dapp
           cli_options[:dry_run]
         end
 
+        def dev_mode?
+          cli_options[:dev].nil? ? config._dev_mode : cli_options[:dev]
+        end
+
         def log_info(*args, **kwargs)
           kwargs[:style] = :info
           log(*args, **kwargs)
