--- lib/dapp/dimg/config/base.rb
@@ -1,7 +1,0 @@
-module Dapp
-  module Dimg
-    module Config
-      class Base < ::Dapp::Config::Base; end
-    end
-  end
-end
