--- lib/dapp/dimg/config/dimg.rb
@@ -1,24 +1,0 @@
-module Dapp
-  module Dimg
-    module Config
-      class Dimg < Base
-        include Validation
-        include InstanceMethods
-
-        attr_reader :_name
-
-        def initialize(name, dapp:)
-          self._name = name
-          super(dapp: dapp)
-        end
-
-        def _name=(name)
-          return if name.nil?
-          reg = '^[[[:alnum:]]_.-]*$'
-          raise Error::Config, code: :dimg_name_incorrect, data: { name: name, reg: reg } unless name =~ /#{reg}/
-          @_name = name
-        end
-      end
-    end
-  end
-end
