+++ lib/dapp/dimg/dapp/command/bp.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Bp
           def bp(repo)
