+++ lib/dapp/dimg/dapp/command/build.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Build
           def build
