+++ lib/dapp/dimg/dapp/command/build_context/export.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module BuildContext
           module Export
