+++ lib/dapp/dimg/dapp/command/list.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module List
           def list
