+++ lib/dapp/dimg/dapp/command/run.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Run
           def run(docker_options, command)
