+++ lib/dapp/dimg/dapp/command/spush.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Spush
           def spush(repo)
