+++ lib/dapp/dimg/dapp/command/stage_image.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module StageImage
           def stage_image
