+++ lib/dapp/dimg/dapp/command/stages/flush_local.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Stages
           module FlushLocal
