+++ lib/dapp/dimg/dapp/command/tag.rb
@@ -1,6 +1,6 @@
 module Dapp
   module Dimg
-    class Dapp
+    module Dapp
       module Command
         module Tag
           def tag(tag)
