+++ lib/dapp/dimg/dapp/dappfile.rb
@@ -1,0 +1,19 @@
+module Dapp
+  module Dimg
+    module Dapp
+      module Dappfile
+        def build_configs
+          @build_configs ||= begin
+            config._dimg.select do |dimg|
+              dimgs_patterns.any? { |pattern| dimg._name.nil? || File.fnmatch(pattern, dimg._name) }
+            end.tap do |dimgs|
+              raise ::Dapp::Error::Dapp, code: :no_such_dimg, data: { dimgs_patterns: dimgs_patterns.join(', ') } if dimgs.empty?
+            end
+          end
+        end
+      end # Dappfile
+    end # Dapp
+  end # Dimg
+end # Dapp
+
+::Dapp::Dapp.send(:include, ::Dapp::Dimg::Dapp::Dappfile)
